web_back/requirements.txt
alembic==1.5.6
blinker==1.4
captcha==0.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-Sockets==0.2.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gevent==20.9.0
gevent-websocket==0.10.1
graphic-verification-code==1.0.3
greenlet==0.4.17
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.4
MarkupSafe==1.1.1
Pillow==7.2.0
PyMySQL==0.10.1
python-dateutil==2.8.1
python-editor==1.0.4
redis==3.5.3
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.20
tqdm==4.50.2
urllib3==1.25.10
websocket==0.2.1
Werkzeug==1.0.1
WTForms==2.3.3
yapf @ file:///tmp/build/80754af9/yapf_1593528177422/work
zope.event==4.5.0
zope.interface==5.1.2
